add saving file ids of tiktok and youtube videos

+++ handlers/youtube.py
@@ -1,16 +1,16 @@
 import asyncio
 import datetime
 import os
-from pytubefix import YouTube
 
 from aiogram import types, Router, F
 from aiogram.types import FSInputFile
 from moviepy.editor import VideoFileClip
+from pytubefix import YouTube
 
-from config import OUTPUT_DIR
-from main import bot, db
-from handlers.user import update_info
 import messages as bm
+from config import OUTPUT_DIR
+from handlers.user import update_info
+from main import bot, db
 
 MAX_FILE_SIZE = 1 * 1024 * 1024
 
@@ -26,6 +26,7 @@ def download_youtube_video(video, name):
 async def download_video(message: types.Message):
     await bot.send_chat_action(message.chat.id, "typing")
     bot_url = f"t.me/{(await bot.get_me()).username}"
+    file_type = "video"
 
     url = message.text
     try:
@@ -43,25 +44,42 @@ async def download_video(message: types.Message):
                 await message.reply("The URL does not seem to be a valid YouTube video link.")
                 return
 
+        post_caption = yt.title
+
+        user_captions = await db.get_user_captions(message.from_user.id)
+
+        db_file_id = await db.get_file_id(yt.watch_url)
+
+        if db_file_id:
+            await bot.send_video(chat_id=message.chat.id, video=db_file_id[0][0],
+                                 caption=bm.captions(user_captions, post_caption, bot_url), parse_mode="HTMl")
+            return
+
         size = video.filesize_kb
 
         if size < MAX_FILE_SIZE:
-            user_captions = await db.get_user_captions(message.from_user.id)
 
             video_file_path = os.path.join(OUTPUT_DIR, name)
             loop = asyncio.get_event_loop()
             await loop.run_in_executor(None, download_youtube_video, video, name)
 
             # Check file size using moviepy
-            post_caption = yt.title
 
             video_clip = VideoFileClip(video_file_path)
 
             width, height = video_clip.size
+
             # Send video file
-            await bot.send_video(chat_id=message.chat.id, video=FSInputFile(video_file_path), width=width,
-                                 height=height,
-                                 caption=bm.captions(user_captions, post_caption, bot_url), parse_mode="HTMl")
+            sent_message = await bot.send_video(chat_id=message.chat.id, video=FSInputFile(video_file_path),
+                                                width=width,
+                                                height=height,
+                                                caption=bm.captions(user_captions, post_caption, bot_url),
+                                                parse_mode="HTMl")
+
+            file_id = sent_message.video.file_id
+
+            await db.add_file(yt.watch_url, file_id, file_type)
+
             os.remove(video_file_path)
 
         else:
@@ -101,26 +119,27 @@ async def download_music(message: types.Message):
             await message.reply("The URL does not seem to be a valid YouTube music link.")
             return
 
+        file_size = audio.filesize_kb
+
+        bitrate_kbps = ''.join(filter(str.isdigit, audio.abr))
+
+        duration_seconds = round((file_size * 8) / int(bitrate_kbps))
+
         audio_file_path = os.path.join(OUTPUT_DIR, name)
 
         loop = asyncio.get_event_loop()
         await loop.run_in_executor(None, download_youtube_video, audio, name)
 
         # Check file size
-        file_size = audio.filesize_kb
-
         if file_size > MAX_FILE_SIZE:
             os.remove(audio_file_path)
             await message.reply("The audio file is too large.")
             return
 
-        bitrate_kbps = ''.join(filter(str.isdigit, audio.abr))
-
-        duration_seconds = round((file_size * 8) / int(bitrate_kbps))
-
         # Send audio file
         await bot.send_audio(chat_id=message.chat.id, audio=FSInputFile(audio_file_path), title=yt.title,
-                             performer=yt.author, duration=duration_seconds, caption=bm.captions(None, None, bot_url),
+                             performer=yt.author, duration=duration_seconds,
+                             caption=bm.captions(None, None, bot_url),
                              parse_mode="HTML")
 
         os.remove(audio_file_path)
